# Optimization: Improve query latency