--- a/lolraft/src/process/thread/query_execution.rs
+++ b/lolraft/src/process/thread/query_execution.rs
@@ -16,9 +16,8 @@ impl Thread {
     fn do_loop(self) -> ThreadHandle {
         let fut = async move {
             loop {
-                self.consumer
-                    .consume_events(Duration::from_millis(100))
-                    .await;
+                // I am not sure the timeout here is necessary.
+                self.consumer.consume_events(Duration::from_secs(1)).await;
                 while self.advance_once().await {
                     tokio::task::yield_now().await;
                 }