before:

```
test query_1    ... bench: 101,825,521.20 ns/iter (+/- 643,273.83)
test query_10   ... bench: 102,293,774.80 ns/iter (+/- 855,290.65)
test query_100  ... bench: 107,019,268.20 ns/iter (+/- 3,038,655.02)
test query_1000 ... bench: 147,752,912.80 ns/iter (+/- 50,666,616.07)
```

after:

```
test query_1    ... bench:     705,891.88 ns/iter (+/- 412,948.82)
test query_10   ... bench:   1,306,312.68 ns/iter (+/- 299,299.63)
test query_100  ... bench:  10,920,550.30 ns/iter (+/- 3,844,038.89)
test query_1000 ... bench: 108,877,229.30 ns/iter (+/- 28,570,455.31)
```